- demo index - removed extra compose2 components - moved upload_endpoint from attachments to AttachmentsUploader - Compose - removed extra Attachment interface
