+++ demo/index.html
@@ -335,7 +335,6 @@
       </temba-tab>
 
       <temba-tab name="compose" icon="compose">
-        Compose
         <div class="example">
           <temba-compose chatbox counter></temba-compose>
         </div>
@@ -350,22 +349,6 @@
         </div>
         <div class="example">
           <temba-compose chatbox attachments counter button></temba-compose>
-        </div>
-        Compose2
-        <div class="example">
-          <temba-compose2 chatbox counter></temba-compose>
-        </div>
-        <div class="example">
-          <temba-compose2 chatbox counter button></temba-compose>
-        </div>
-        <div class="example">
-          <temba-compose2 attachments></temba-compose>
-        </div>
-        <div class="example">
-          <temba-compose2 attachments button></temba-compose>
-        </div>
-        <div class="example">
-          <temba-compose2 chatbox attachments counter button></temba-compose>
         </div>
       </temba-tab>
